correct  user deletion

+++ app.py
@@ -183,19 +183,23 @@ def add_user():
 
 
 @app.route('/delete/<int:id>')
+@login_required
 def delete(id):
-    name_to_delete = Users.query.get_or_404(id)
-    name = None
-    form = UserForm()
-    try:
-        db.session.delete(name_to_delete)
-        db.session.commit()
-        flash("User Deleted successfully")
-        our_users = Users.query.order_by(Users.date_added)
-        return render_template("add_user.html", form=form, name=name, our_users=our_users)
-    except:
-        flash("There was a problem deleting this record... Please try again later!")
-
+    if id == current_user.id:
+        name_to_delete = Users.query.get_or_404(id)
+        name = None
+        form = UserForm()
+        try:
+            db.session.delete(name_to_delete)
+            db.session.commit()
+            flash("User Deleted successfully")
+            our_users = Users.query.order_by(Users.date_added)
+            return render_template("add_user.html", form=form, name=name, our_users=our_users)
+        except:
+            flash("There was a problem deleting this record... Please try again later!")
+    else:
+        flash("Sorry, you cant delete that user....")
+        return redirect(url_for('dashboard'))
 
 @app.route('/')
 def index():
